fix: changed administrate api
After saving an event an error will show up. This error happens on the
show page of the event. It seems like the administrate api has changed
and the attributes are now nested inside a hash with no given key.

This commit returns the correct attributes for the show page and
resolves the attributes correctly.