Improve texts if user group has no twitter account
Some texts assumed that every user group has a twitter account which
might sometimes not be the case. For example the dresden.rb user group
does not have a twitter account (at least not yet). This change removes
the "requirement" and improves the site if no twitter account was given.