Use next event as the "current" event
This change removes the restriction of the current event to be at most 9
weeks in the future. Some user groups might decide to hold meetups less
frequently so this limitation prevents them from advertising their next
event until 9 weeks before the meetup.

With this change the current event that is displayed on the homepage
will always be the next event, even if it's set far in the future.

Closes #1084